requirements.txt
# --- Core Web Framework ---
fastapi                  # High-performance API framework
uvicorn[standard]        # ASGI web server
python-multipart         # Required for OAuth2 form data parsing

# --- Database & ORM ---
sqlalchemy               # Core database toolkit and ORM
pymysql                  # MySQL database driver
alembic                  # Database schema migration tool

# --- Security & Authentication ---
passlib[bcrypt]          # Cryptographic password hashing
python-jose[cryptography] # JWT creation and validation (REQUIRED for oauth2.py)

# --- Utilities & Validation ---
pydantic                 # Data validation and serialization
pydantic-settings        # Environment variable management
python-dotenv            # Local .env file parsing
email-validator          # Strict email format validation
resend                   # Transactional email API (REQUIRED for utils.py)